backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
google-generativeai
python-multipart
cryptography>=41.0.0
